% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_connect_layer.R
\name{build_connect_layer}
\alias{build_connect_layer}
\title{Create raster using passage and resistance layer}
\usage{
build_connect_layer(
  predRasterLoc,
  patchList,
  akdeSummary,
  REGION,
  prelimAggFact = NA,
  seed = 2025,
  THETA = NULL,
  repeatsPerPair,
  MSEdf = NULL,
  MINPATCHSIZE,
  cropArea = 750,
  cores = 4
)
}
\value{
abc
}
\description{
abc
}
