% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predResistance_layer.R
\name{build_predResistance_layer}
\alias{build_predResistance_layer}
\title{Create raster based on the mean SSF coefs}
\usage{
build_predResistance_layer(
  ssfData,
  ssfPoismodel,
  landuseList,
  patchList,
  deerData,
  REGION,
  prelimAggFact,
  inlaFormula = NULL
)
}
\value{
abc
}
\description{
abc
}
