% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predResistanceFallow_layer.R
\name{build_predResistanceFallow_layer}
\alias{build_predResistanceFallow_layer}
\title{Create raster based on the mean SSF coefs}
\usage{
build_predResistanceFallow_layer(
  ssfData,
  ssfPoismodel,
  projLayer,
  prelimAggFact,
  inlaFormula = NULL
)
}
\value{
abc
}
\description{
abc
}
