% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predResistanceRodent_layer.R
\name{build_predResistanceRodent_layer}
\alias{build_predResistanceRodent_layer}
\title{Create raster based on the mean SSF coefs}
\usage{
build_predResistanceRodent_layer(projLayer, prelimAggFact)
}
\value{
abc
}
\description{
abc
}
