% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dbbmm_mse.R
\name{calculate_dbbmm_mse}
\alias{calculate_dbbmm_mse}
\title{Compare connectivity layer with all dBBMMs}
\usage{
calculate_dbbmm_mse(deerData, dbbmmList, connectRasterLoc, REGION, THETA)
}
\value{
abc
}
\description{
abc
}
