% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dbbmms.R
\name{calculate_dbbmms}
\alias{calculate_dbbmms}
\title{Create dbbmms to test against distance maps}
\usage{
calculate_dbbmms(deerData, landuseList, window, margin, locationError)
}
\value{
abc
}
\description{
abc
}
