% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package_txt.R
\name{create_package_txt}
\alias{create_package_txt}
\title{Use grateful to generate a paragraph of R packages}
\usage{
create_package_txt(excludes = NULL, rstudioVersion = "2024.12.0+467")
}
\arguments{
\item{excludes}{character string of packages to exclude from the paragraph.}
}
\value{
A txt file output and returns that string of text.
}
\description{
Uses grateful to generate a paragraph of R packages that can be used in a RMD document.
}
