% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_psuedo_abs.R
\name{create_psuedo_abs}
\alias{create_psuedo_abs}
\title{Build a biased pseudo-absence dataset}
\usage{
create_psuedo_abs(
  occData,
  hfBiasLayer = here::here("data", "Human Footprint", "hfp2022.tif"),
  nPointMultiplier = 3,
  nReps = 3,
  envLayers = read_stack_layers(layerLoc = here("data", "GIS data", "SDM Layers"))
)
}
\value{
A bm_PseudoAbsences object
}
\description{
Using human footprint
}
