% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_rodent_HR.R
\name{determine_rodent_HR}
\alias{determine_rodent_HR}
\title{Read HR database and summarise}
\usage{
determine_rodent_HR()
}
\value{
A value and two tables summarising studies effort
}
\description{
a
}
