% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_akde_summaries.R
\name{extract_akde_summaries}
\alias{extract_akde_summaries}
\title{Extract summary stats from AKDE and complete overall mean for Aberdeen deer}
\usage{
extract_akde_summaries(deerData, akdeLists)
}
\value{
a
}
\description{
a
}
