% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_akdeFallow_longest.R
\name{extract_akdeFallow_longest}
\alias{extract_akdeFallow_longest}
\title{Extract longest Axis Summary for fallow deer}
\usage{
extract_akdeFallow_longest(akdeLists)
}
\value{
a
}
\description{
a
}
