% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_connectivity_locations.R
\name{extract_connectivity_locations}
\alias{extract_connectivity_locations}
\title{Extract data from connectivity layers to compare to movement data}
\usage{
extract_connectivity_locations(
  connectRasterLocations,
  connectRasterLocationWessex,
  akdeLists,
  deerData,
  MSEdf,
  conAvail,
  nAvail,
  typeAvial,
  seed = 2025
)
}
\value{
A list of available and used dataframes.
}
\description{
Extract data from connectivity maps
}
