% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_patch_connectivity.R,
%   R/summarise_patch_connectivity.R
\name{extract_patch_connectivity}
\alias{extract_patch_connectivity}
\title{Summarise Patch Connectivity}
\usage{
extract_patch_connectivity(
  MSEdf = NULL,
  connectRasterLocations,
  patchList,
  selectedPatchList,
  REGION,
  buffers
)

extract_patch_connectivity(
  MSEdf = NULL,
  connectRasterLocations,
  patchList,
  selectedPatchList,
  REGION,
  buffers
)
}
\value{
abc

abc
}
\description{
abc

abc
}
