% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_patch_omniConnectivity.R
\name{extract_patch_omniConnectivity}
\alias{extract_patch_omniConnectivity}
\title{Summarise Patch Connectivity}
\usage{
extract_patch_omniConnectivity(omniNormCurr, selectedPatchList, buffers)
}
\value{
abc
}
\description{
abc
}
