% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rsf_results.R
\name{extract_rsf_results}
\alias{extract_rsf_results}
\title{Extract results and plot rsf outputs}
\usage{
extract_rsf_results(deerData, rsfDataList, rsfModelList, nAvail)
}
\value{
Multiple plots and a table of model metrics.
}
\description{
Check models and plot outputs from individual RSF models.
}
