% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overview_maps.R
\name{generate_overview_maps}
\alias{generate_overview_maps}
\title{Generate overview maps per deer}
\usage{
generate_overview_maps(deerData, akdeLists, landuseList, patches)
}
\value{
A list of ggplot object, and a saved file per individual.
}
\description{
AKDE outputs mapped against land use and patches.
}
