% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palettes.R
\name{generate_palettes}
\alias{generate_palettes}
\title{Generate palettes for plots}
\usage{
generate_palettes()
}
\value{
List of various named or DF colour schemes
}
\description{
Pre-created stored palettes to keep colours consistent
}
