% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tracking_plots.R
\name{generate_tracking_plots}
\alias{generate_tracking_plots}
\title{Generate tracking summary plots}
\usage{
generate_tracking_plots(deerData)
}
\value{
A two object list, one object for duration plot one for timelag.
}
\description{
Take all deer data and generate plots of tracking duration and timelag.
}
