% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tracking_table.R
\name{generate_tracking_table}
\alias{generate_tracking_table}
\title{Generate summary of tracking data}
\usage{
generate_tracking_table(deerData)
}
\value{
A dataframe of summarised tracking data. Also exports a csv table.
}
\description{
Simple summary of tracking frequency and duration.
}
