% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_connectivity_thetas.R
\name{map_connectivity_thetas}
\alias{map_connectivity_thetas}
\title{Map out theta variation}
\usage{
map_connectivity_thetas(connectRasterLocations, landuseList, patchList, REGION)
}
\value{
abc
}
\description{
abc
}
