% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity_distributions.R
\name{plot_connectivity_distributions}
\alias{plot_connectivity_distributions}
\title{Create a plot of connectivity values at different areas}
\usage{
plot_connectivity_distributions(validData, connectivityValue_models)
}
\value{
abc
}
\description{
abc
}
