% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_to_patch.R
\name{plot_distance_to_patch}
\alias{plot_distance_to_patch}
\title{Generate plot to explore distance from patch}
\usage{
plot_distance_to_patch(deerData, patchList, akdeSummary)
}
\value{
abc
}
\description{
abc
}
