% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcStruc_comparison.R
\name{plot_funcStruc_comparison}
\alias{plot_funcStruc_comparison}
\title{Plot connectivity values against structural connectivity}
\usage{
plot_funcStruc_comparison(patchData)
}
\value{
abc
}
\description{
abc
}
