% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_homeRange_sizes.R
\name{plot_homeRange_sizes}
\alias{plot_homeRange_sizes}
\title{Generate plot showing home range sizes}
\usage{
plot_homeRange_sizes(deerData, akdeLists)
}
\value{
A ggplot object, and saved file.
}
\description{
Plot all row deer home range estimates.
}
