% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occSDMOmni_inOut.R
\name{plot_occSDMOmni_inOut}
\alias{plot_occSDMOmni_inOut}
\alias{plot_occSDMOmni_inOut_HFOcc}
\title{Create a plot all aspects of omni pipeline}
\usage{
plot_occSDMOmni_inOut_HFOcc(
  species,
  hfBiasLayer = here("data", "Human Footprint", "hfp2022.tif"),
  pseudoAbs,
  ...
)
}
\value{
abc
}
\description{
abc
}
