% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patch_connectivity.R
\name{plot_patch_connectivity}
\alias{plot_patch_connectivity}
\title{Map out patch summaries}
\usage{
plot_patch_connectivity(
  MSEdf,
  connectRasterLocations,
  patchList,
  selectedPatchList,
  REGION,
  bufferSummaries
)
}
\value{
abc
}
\description{
abc
}
