% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patch_MMMconnectivity.R
\name{plot_patch_MMMconnectivity}
\alias{plot_patch_MMMconnectivity}
\title{Map out patch summaries}
\usage{
plot_patch_MMMconnectivity(
  MSEdf,
  connectRasterLocations,
  patchList,
  selectedPatchList,
  REGION
)
}
\value{
abc
}
\description{
abc
}
