% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patch_summary.R
\name{plot_patch_summary}
\alias{plot_patch_summary}
\title{Plot patch connectivity}
\usage{
plot_patch_summary(
  bufferSummaries,
  MSEdf,
  connectRasterLocations,
  patchList,
  selectedPatchList,
  REGION
)
}
\value{
abc
}
\description{
abc
}
