% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_circuitscape_data.R
\name{prepare_circuitscape_data}
\alias{prepare_circuitscape_data}
\title{Create rasters suitable for the circuitscape}
\usage{
prepare_circuitscape_data(
  predRasterLoc,
  patchList,
  REGION,
  prelimAggFact = 10,
  patchDistance
)
}
\value{
abc
}
\description{
abc
}
