% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rsf_data.R
\name{prepare_rsf_data}
\alias{prepare_rsf_data}
\title{Extract data from landscape layers for RSF}
\usage{
prepare_rsf_data(
  deerData,
  akdeLists,
  landuseList,
  patchList,
  nAvail = 10,
  typeAvial = "random",
  conAvail = "99\%"
)
}
\value{
A list of available and used dataframes.
}
\description{
Extract data from landscape ready for RSF models.
}
