% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sdm_layers.R
\name{prepare_sdm_layer}
\alias{prepare_sdm_layer}
\title{Read land use and other land data and make rasters}
\usage{
prepare_sdm_layer(prelimAggFact = NULL)
}
\value{
A list of raster file locations
}
\description{
Reads in the UKCEH land use data
}
