% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ssfFallow_data.R
\name{prepare_ssfFallow_data}
\alias{prepare_ssfFallow_data}
\title{Extract data from landscape layers for SSF}
\usage{
prepare_ssfFallow_data(
  deerData,
  landuseList,
  patchList,
  cores = 8,
  nAvail = 10,
  slDist = "gamma",
  taDist = "vonmises"
)
}
\value{
A list of available and used dataframes.
}
\description{
Extract data from landscape ready for SSF models.
}
