% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cleanFallow_occData.R
\name{read_cleanFallow_occData}
\alias{read_cleanFallow_occData}
\title{Read fallow occ data}
\usage{
read_cleanFallow_occData(sdmLayers)
}
\value{
A sf occ data
}
\description{
Read in a clean occ data for UK fallow
}
