% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cleanRodent_occData.R
\name{read_cleanRodent_occData}
\alias{read_cleanRodent_occData}
\title{Read rodent occ data}
\usage{
read_cleanRodent_occData(sdmLayers)
}
\value{
A sf occ data
}
\description{
Read in a clean occ data for UK rodents
}
