% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_akdes.R, R/read_deer_data.R,
%   R/read_landuse_data.R
\name{read_deer_data}
\alias{read_deer_data}
\alias{calculate_akdes}
\alias{read_landuse_data}
\title{Create variograms, choose best fitting model, and created weighted AKDE}
\usage{
calculate_akdes(deerData)

read_deer_data()

read_landuse_data(deerData, patchList, prelimAggFact)
}
\value{
abc

A dataframe with all the deer data processed, and an exported processed csv for review.

A list of Aberdeen and Wessex land data.
}
\description{
abc

A function to read in raw data and do basic clean up and add converted 27700 coord system.

Reads in the UKCEH land use data for the two areas and also OS road data.
}
