% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_patches_data.R
\name{read_patches_data}
\alias{read_patches_data}
\title{Read forest patche spatial data}
\usage{
read_patches_data()
}
\value{
A two object list, one object for each study location.
}
\description{
Read in the previously created patch information.
}
