% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_julia_circuitscape.R
\name{run_julia_circuitscape}
\alias{run_julia_circuitscape}
\title{Julia call wrapper to run circuitscape}
\usage{
run_julia_circuitscape(model, circuitScapeData)
}
\value{
abc
}
\description{
abc
}
