% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pois_model.R
\name{run_pois_model}
\alias{run_pois_model}
\title{Run Poisson model for overall selection}
\usage{
run_pois_model(ssfDataList)
}
\value{
The INLA output.
}
\description{
Run the Muff et al poisson model.
}
