% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rsf_models.R
\name{run_rsf_models}
\alias{run_rsf_models}
\title{Run basic RSFs}
\usage{
run_rsf_models(rsfDataList, rsfFormula)
}
\value{
A list of model outputs from the glmer function.
}
\description{
Run basic RSF using glmer.
}
