% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_connect_layer.R
\name{standardise_connect_layer}
\alias{standardise_connect_layer}
\title{Set connect layers between 0 and 1}
\usage{
standardise_connect_layer(connectRasterLoc, REGION, THETA = NULL, MSEdf = NULL)
}
\value{
abc
}
\description{
abc
}
