% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_validation_data.R
\name{test_validation_data}
\alias{test_validation_data}
\title{Run logisitic regressions on real data to compare to connectivity values}
\usage{
test_validation_data(validData)
}
\value{
abc
}
\description{
abc
}
